HSEngine/requirements.txt
beautifultable==1.0.0
fate-client>=0.3.0
gmpy2==2.0.8
joblib>=1.0.1
kfserving==0.5.1
kubernetes==12.0.1
minio>=6.0.2
numpy==1.19.5
pycryptodomex==3.6.6
pytorch-lightning>=1.2.1
ruamel.yaml>=0.16.0
scikit-learn==0.19.2
scipy==1.1.0
torch-model-archiver>=0.3.1
torchvision==0.5.0